fix unsigned-signed comparison

+++ jni/network.cpp
@@ -671,7 +671,7 @@ void net::udp::send(const void *buffer,unsigned len){
 
 	// no such thing as a partial send for udp with sendto
 	ssize_t result=sendto(sock,(const char*)buffer,len,0,(sockaddr*)ai->ai_addr,ai->ai_addrlen);
-	if(result!=len){
+	if((unsigned)result!=len){
 		this->close();
 		return;
 	}
@@ -687,7 +687,7 @@ void net::udp::recv(void *buffer,unsigned len){
 
 	// no such thing as a partial send for udp with sendto
 	ssize_t result=recvfrom(sock,(char*)buffer,len,0,(sockaddr*)&src_addr,&src_len);
-	if(result!=len){
+	if((unsigned)result!=len){
 		this->close();
 		return;
 	}
